Improve handling of DOM Nodes
This avoids race conditions that sometimes occurred in the
willReceiveProps callback when a timeout caused findDOMNode to be called
on an instance that had since been unmounted.

See https://github.com/zippyui/react-datagrid/issues/89

The fix uses react-0.14 ref-callbacks to store references to the DOM
nodes, when we _know_ they exist, instead of trying to obtain the
references at random times.